requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Flask==1.0.2
gitdb2==2.0.5
GitPython==2.1.11
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
netaddr==0.7.19
requests==2.21.0
smmap2==2.0.5
urllib3==1.24.1
Werkzeug==0.14.1
